+++ android/app/src/main/kotlin/za/co/towntrek/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.towntrek_flutter
+package za.co.towntrek.app
 
 import io.flutter.embedding.android.FlutterActivity