+++ lib/repositories/repositories.dart
@@ -3,3 +3,4 @@ export 'business_repository.dart';
 export 'town_repository.dart';
 export 'event_repository.dart';
 export 'service_repository.dart';
+export 'stats_repository.dart';